Move existing handler into Core
